Separate redeploy docker stack command
